+++ R/layout_matrix.R
@@ -8,6 +8,17 @@
 #' This function allows user to create the matrix using a list of
 #' coordinates for the node labels.
 #'
+#' Note that [semPlot::semPaths()] only
+#' treat a matrix as a layout matrix as
+#' a matrix of grid if it has three or
+#' more columns. Therefore, if the resulting
+#' matrix has only one or two columns, it
+#' will be converted to a matrix with
+#' three columns, by inserting a column
+#' of `NA`s for a two-column matrix, and
+#' inserting the resulting matrix between
+#' two columns of `NA`s for a one-column
+#' matrix.
 #'
 #'@return A layout matrix for the layout argument of
 #'[semPlot::semPaths()].
@@ -45,5 +56,13 @@ layout_matrix <- function(...) {
       jname <- names(layout)[i]
       out[j[1], j[2]] <- jname
     }
+  # semPlot::semPaths() treat a matrix as
+  # a layout matrix only if it has three
+  # or more columns.
+  if (ncol(out) == 2) {
+    out <- cbind(out[, 1], NA, out[, 2])
+  } else if (ncol(out) == 1) {
+    out <- cbind(NA, out[, 1], NA)
+  }
   out
   }