+++ tests/testthat/test-set_edge_attribute.R
@@ -34,6 +34,34 @@ p1 <- set_edge_attribute(p_pa, c("x1 ~~x2" = "red",
 
 # plot(p1)
 
+# An edge not in the model will be skipped
+
+p2b <- set_edge_attribute(p_pa, c("x1 ~~x2" = "red",
+                                 "x4~~ x3" = "black",
+                                 "x3 ~ x1" = "white",
+                                 "x1 ~ x4" = "blue",
+                                 "x4 ~ x2" = "yellow"),
+                         attribute_name = "color")
+# plot(p2b)
+
+p2c <- set_edge_attribute(p_pa, c("x1 ~~x2" = "red",
+                                 "x4~~ x3" = "black",
+                                 "x3 ~ x1" = "white",
+                                 "x1 ~~ x4" = "blue",
+                                 "x4 ~ x2" = "yellow"),
+                         attribute_name = "color")
+# plot(p2c)
+
+# Direction ignored
+p2d <- set_edge_attribute(p_pa, c("x1 ~~x2" = "red",
+                                 "x4~~ x3" = "black",
+                                 "x3 ~~ x1" = "white",
+                                 "x4 ~~ x1" = "darkgreen",
+                                 "x4 ~~ x2" = "yellow"),
+                         attribute_name = "color",
+                         check_direction = FALSE)
+# plot(p2d)
+
 test_that("set_edge_attribute: color", {
     expect_equal(p2$graphAttributes$Edges$color,
                  c("blue", "#808080FF", "#808080FF", "#808080FF", "#808080FF",
@@ -42,6 +70,14 @@ test_that("set_edge_attribute: color", {
     expect_equal(p1$graphAttributes$Edges$color,
                  c("red", "white", "#808080FF", "darkgreen", "yellow", "#808080FF",
                    "#808080FF", "#808080FF", "#808080FF", "black", "red", "black"))
+    expect_equal(p2b$graphAttributes$Edges$color,
+                 c("red", "white", "#808080FF", "#808080FF", "yellow", "#808080FF",
+                   "#808080FF", "#808080FF", "#808080FF", "black", "red", "black"))
+    expect_equal(p2c$graphAttributes$Edges$color,
+                 c("red", "white", "#808080FF", "#808080FF", "yellow", "#808080FF",
+                   "#808080FF", "#808080FF", "#808080FF", "black", "red", "black"))
+    expect_equal(p2d$graphAttributes$Edges$color,
+                 p1$graphAttributes$Edges$color)
   })
 
 p3 <- set_edge_attribute(p_pa, c("x1 ~~x2" = -3,
@@ -64,4 +100,16 @@ p4 <- set_edge_attribute(p_pa, c("x1 ~~x2" = 2,
 test_that("set_edge_attribute: label.cex", {
     expect_equal(p4$graphAttributes$Edges$label.cex,
                  c(2, 5, 1.15, 1.15, 1.15, 1.15, 1.15, 1.15, 1.15, 1.15, 2, 1.15))
-  })
+})
+
+# Sell all edges
+
+p5 <- set_edge_attribute(p_pa,
+                         5,
+                         attribute_name = "label.cex")
+
+test_that("set_edge_attribute: label.cex", {
+    expect_equal(p5$graphAttributes$Edges$label.cex,
+                 rep(5, 12))
+})
+